Smart Resume Ranker/requirements.txt
flask
pandas
numpy
scikit-learn
matplotlib
fuzzywuzzy
python-docx
PyPDF2
psutil
python-Levenshtein
